refactor: remove venue entity, fix i32→i64 deserialization, update folder workflows
Remove the separate Venue entity — venue data stays as fields on Project
(city, country, area). This simplifies the domain model while preserving
all location information.

Key changes:
- Remove venues.rs commands, VenueModal, Venues page, venue API/stores
- Fix i32→i64 in entities.rs for SurrealDB v3 binary protocol compat
- Update folder status mappings to new domain model (Lead, RFP, Submitted,
  Awarded, Design, Construction, Practical Completion, Completed,
  Superseded, Cancelled, Lost, No Response, On Hold)
- Broaden template copy trigger: fires on ANY move to "11 Current",
  not just from "01 RFPs"
- Update move_project_from_rfp() to use "awarded" instead of "active"
- Rename Accepted→Awarded, Rejected→Lost in frontend status references

+++ src-tauri/src/db/entities.rs
@@ -91,7 +91,7 @@ pub struct ProjectNumber {
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Revision {
-    pub revision_number: i32,
+    pub revision_number: i64,
     pub revision_date: Datetime,
     pub author_email: String,
     pub author_name: String,
@@ -144,7 +144,7 @@ pub struct Fee {
     pub staff_phone: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub staff_position: Option<String>,
-    pub rev: i32, // auto-computed
+    pub rev: i64, // auto-computed
     #[serde(default)]
     pub revisions: Vec<Revision>,
     // Pricing data (stored as opaque JSON, passed through to frontend)
@@ -248,7 +248,7 @@ pub struct FeeCreate {
     pub staff_phone: Option<String>,
     pub staff_position: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub rev: Option<i32>,
+    pub rev: Option<i64>,
     #[serde(default)]
     pub revisions: Vec<Revision>,
 }
@@ -269,7 +269,7 @@ pub struct FeeUpdate {
     pub staff_email: Option<String>,
     pub staff_phone: Option<String>,
     pub staff_position: Option<String>,
-    pub rev: i32,
+    pub rev: i64,
     #[serde(default)]
     pub revisions: Vec<Revision>,
 }
